feat: lockdown mutations for authenticated users only

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -57,8 +57,9 @@ const Mutation = {
 		}
 	},
 
-	async updateUser(parent, args, { prisma }, info) {
-		const userExists = await prisma.exists.User({ id: args.id })
+	async updateUser(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const userExists = await prisma.exists.User({ id: userId })
 
 		if (!userExists) {
 			throw new Error('User not found')
@@ -67,7 +68,7 @@ const Mutation = {
 		return prisma.mutation.updateUser(
 			{
 				where: {
-					id: args.id,
+					id: userId,
 				},
 				data: args.data,
 			},
@@ -75,14 +76,15 @@ const Mutation = {
 		)
 	},
 
-	async deleteUser(parent, args, { prisma }, info) {
-		const userExists = await prisma.exists.User({ id: args.id })
+	async deleteUser(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const userExists = await prisma.exists.User({ id: userId })
 
 		if (!userExists) {
 			throw new Error('User not found')
 		}
 
-		return prisma.mutation.deleteUser({ where: { id: args.id } }, info)
+		return prisma.mutation.deleteUser({ where: { id: userId } }, info)
 	},
 
 	async createPost(parent, args, { prisma, request }, info) {
@@ -109,11 +111,16 @@ const Mutation = {
 		)
 	},
 
-	async updatePost(parent, args, { prisma }, info) {
-		const postExists = await prisma.exists.Post({ id: args.id })
+	async updatePost(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+
+		const postExists = await prisma.exists.Post({
+			id: args.id,
+			author: { id: userId },
+		})
 
 		if (!postExists) {
-			throw new Error('Post not found')
+			throw new Error('Unable to update post')
 		}
 
 		return prisma.mutation.updatePost(
@@ -127,24 +134,31 @@ const Mutation = {
 		)
 	},
 
-	async deletePost(parent, args, { prisma }, info) {
-		const postExists = await prisma.exists.Post({ id: args.id })
+	async deletePost(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+
+		const postExists = await prisma.exists.Post({
+			id: args.id,
+			author: { id: userId },
+		})
 
 		if (!postExists) {
-			throw new Error('Post not found')
+			throw new Error('Unable to delete Post')
 		}
 
 		return prisma.mutation.deletePost({ where: { id: args.id } }, info)
 	},
 
-	async createComment(parent, args, { prisma }, info) {
+	async createComment(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+
 		return prisma.mutation.createComment(
 			{
 				data: {
 					text: args.data.text,
 					author: {
 						connect: {
-							id: args.data.author,
+							id: userId,
 						},
 					},
 					post: {
@@ -158,7 +172,18 @@ const Mutation = {
 		)
 	},
 
-	async updateComment(parent, args, { prisma }, info) {
+	async updateComment(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+
+		const commentExists = await prisma.exists.Comment({
+			id: args.id,
+			author: { id: userId },
+		})
+
+		if (!commentExists) {
+			throw new Error('Unable to update comment')
+		}
+
 		return prisma.mutation.updateComment(
 			{
 				where: {
@@ -170,7 +195,18 @@ const Mutation = {
 		)
 	},
 
-	async deleteComment(parent, args, { prisma }, info) {
+	async deleteComment(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+
+		const commentExists = await prisma.exists.Comment({
+			id: args.id,
+			author: { id: userId },
+		})
+
+		if (!commentExists) {
+			throw new Error('Unable to update comment')
+		}
+
 		return prisma.mutation.deleteComment(
 			{
 				where: {
